requirements.txt
absl-py==0.13.0
albumentations==1.0.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
efficientnet-pytorch==0.7.1
einops==0.3.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
grpcio==1.38.1
idna==2.10
imageio==2.9.0
imgaug==0.4.0
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1618853849286/work
mkl-service==2.3.0
networkx==2.5.1
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1620830962040/work
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.2.54
opencv-python-headless==4.5.2.54
pandas==1.3.0
Pillow @ file:///tmp/build/80754af9/pillow_1625655817137/work
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
Shapely==1.7.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.2.0
tifffile==2021.7.2
timm==0.4.12
torch==1.8.0
torchvision==0.9.0
tqdm==4.61.2
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
urllib3==1.26.6
Werkzeug==2.0.1
